requirements.txt
feedparser>=6.0.10
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
schedule>=1.2.1
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Bcrypt>=1.0.1
requests>=2.31.0
gunicorn>=21.2.0